refactor: move elderly school APIs to /api/elderly-school/*
- pages/api/smart-health/elderly/* (8 ไฟล์) → pages/api/elderly-school/*
- elderly-school-dashboard.js → elderly-school/sheet-dashboard.js
- อัปเดต fetch ทุกจุด (ElderlySchoolDashboard 13 จุด, elderly-cards,
  elderly-schedule, /elderly/checkin)
- คง shim ที่ /api/smart-health/elderly/checkin หนึ่ง release กัน tab ค้าง
  ช่วง deploy (ลบเฟสถัดไป) — path หน้า /elderly/checkin ไม่เปลี่ยน QR เดิมใช้ได้
